Added jwt and statements for post put and delete added comments and api response codes

+++ part3/hbnb/app/api/v1/reviews.py
@@ -1,5 +1,6 @@
 from flask_restx import Namespace, Resource, fields
 from app.services import facade
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 api = Namespace('reviews', description='Review operations')
 
@@ -16,10 +17,24 @@ class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @api.response(404, 'Place not found')
+    @jwt_required
     def post(self):
         """Register a new review"""
-        # Placeholder for the logic to register a new review
         review = api.payload
+        place = facade.get_place(review['place_id'])
+        if not place:
+            return {'error': 'Place not found'}, 404
+        # users can only review places they do not own
+        current_user = get_jwt_identity()
+        if place.owner == current_user['id']:
+            return {'error': 'You cannot review your own place'}
+        # users can only create one review per place
+        place_reviews = facade.get_reviews_by_place(review['place_id'])
+        for existing_review in place_reviews:
+            if existing_review.user_id == current_user['id']:
+                return {'error': 'You have already reviewed this place'}, 400
+
         try:    
             new_review = facade.create_review(review)
             return {'id': new_review.id, 'comment': new_review.comment, 'rating': new_review.rating, 'user_id': new_review.user_id, 'place_id': new_review.place_id}, 201
@@ -28,9 +43,9 @@ class ReviewList(Resource):
 
 
     @api.response(200, 'List of reviews retrieved successfully')
+    @api.response(404, 'No reviews found')
     def get(self):
         """Retrieve a list of all reviews"""
-        # Placeholder for logic to return a list of all reviews
         reviews = facade.get_all_reviews()
         if not reviews:
             return {'error': 'No reviews found'}, 404
@@ -42,7 +57,6 @@ class ReviewResource(Resource):
     @api.response(404, 'Review not found')
     def get(self, review_id):
         """Get review details by ID"""
-        # Placeholder for the logic to retrieve a review by ID
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
@@ -50,24 +64,59 @@ class ReviewResource(Resource):
 
     @api.expect(review_model)
     @api.response(200, 'Review updated successfully')
-    @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @api.response(404, 'Review not found')
+    @api.response(403, 'Unauthorized action')
+    @api.response(500, 'Failed to update this review')
+    @jwt_required()
     def put(self, review_id):
-        """Update a review's information"""
-        # Placeholder for the logic to update a review by ID
+        """Update a review's information by ID"""
         review = facade.get_review(review_id)
+        # check if review exists in the database
         if not review:
             return {'error': 'Review not found'}, 404
+        
+        current_user = get_jwt_identity()
+        # users can only modify reviews they created
+        if review.user_id != current_user['id']:
+            return {'error': 'Unauthorized action'}, 403
+        
+        review_data = api.payload
+        # reviews must have a comment and a rating field
+        if 'comment' not in review_data or 'rating' not in review_data:
+            return {'error': 'Invalid input data'}, 400
+        review.comment = review_data.get('comment', review.comment)
+        review.rating = review_data.get('rating', review.rating)
+        # update review in the database
+        updated_review = facade.update_review(review.id, {
+            'comment': review.comment,
+            'rating': review.rating,
+        })
+        # check if database issue occured when updating review
+        if not updated_review:
+            return {'error': 'Failed to update this review'}, 500
         return {'id': review.id, 'comment': review.comment, 'rating': review.rating, 'user_id': review.user_id, 'place_id': review.place_id}, 200
 
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @api.response(403, 'Unauthorized action')
+    @api.response(500, 'Failed to delete this review')
+    @jwt_required()
     def delete(self, review_id):
         """Delete a review"""
-        # Placeholder for the logic to delete a review
-        review = facade.delete_review(review_id)
+        review = facade.get_review(review_id)
+        # check if review exists in the database
         if not review:
             return {'error': 'Review not found'}, 404
+        # users can only delete reviews they created
+        current_user = get_jwt_identity()
+        if review.user_id != current_user['id']:
+            return {'error': 'Unauthorized action'}, 403
+        # delete review in the database
+        deleted_review = facade.delete_review(review_id)
+        # check if database issue occured when deleting review
+        if not deleted_review:
+            return {'error': 'Failed to delete this review'}, 500
         return {'message': 'Review deleted successfully'}, 200
 
 @api.route('/places/<place_id>/reviews')
@@ -76,7 +125,6 @@ class PlaceReviewList(Resource):
     @api.response(404, 'Place not found')
     def get(self, place_id):
         """Get all reviews for a specific place"""
-        # Placeholder for logic to return a list of reviews for a place
         reviews = facade.get_reviews_by_place(place_id)
         if not reviews:
             return {'error': 'No reviews found for this place'}, 404
